// partners/packages/affiliate-program-api/api.ts
    export enum TypeEnum {
        REVSHARE = <any>"REVSHARE",
        CPA = <any>"CPA",
    }
// ...
export interface AffiliateProgramRegisterRequest {
    /**
     *
     * @type {string}
     * @memberof AffiliateProgramRegisterRequest
     */
    type: AffiliateProgramRegisterRequest.TypeEnum;
    /**
     *
     * @type {string}
     * @memberof AffiliateProgramRegisterRequest
     */
    name: string;
    /**
     *
     * @type {string}
     * @memberof AffiliateProgramRegisterRequest
     */
    email: string;
    /**
     *
     * @type {string}
     * @memberof AffiliateProgramRegisterRequest
     */
    phone_number: string;
    /**
     *
     * @type {string}
     * @memberof AffiliateProgramRegisterRequest
     */
    telegram: string;
    /**
     *
     * @type {string}
     * @memberof AffiliateProgramRegisterRequest
     */
    whatsapp: string;
    /**
     *
     * @type {string}
     * @memberof AffiliateProgramRegisterRequest
     */
    skype: string;
    /**
     *
     * @type {string}
     * @memberof AffiliateProgramRegisterRequest
     */
    password: string;
    /**
     *
     * @type {string}
     * @memberof AffiliateProgramRegisterRequest
     */
    traffic_source: string;
    /**
     *
     * @type {string}
     * @memberof AffiliateProgramRegisterRequest
     */
    currency: string;
}

/**
 * @export
 * @namespace AffiliateProgramRegisterRequest
 */
export namespace AffiliateProgramRegisterRequest {
    /**
     * @export
     * @enum {string}
     */
    export enum TypeEnum {
        REVSHARE = <any>"REVSHARE",
        CPA = <any>"CPA",
    }
}